+++ resources/views/livewire/buttons.blade.php
@@ -1,1 +1,0 @@
-@include('filament-socialite::components.buttons')