+++ tests/TestCase.php
@@ -29,10 +29,5 @@ class TestCase extends Orchestra
     public function getEnvironmentSetUp($app)
     {
         config()->set('database.default', 'testing');
-
-        /*
-        $migration = include __DIR__.'/../database/migrations/create_filament-socialite_table.php.stub';
-        $migration->up();
-        */
     }
 }